Created separate classes and completed function to add items to an order and to delete order
