+++ src/main/kotlin/com/rafeng/bakery/improve/business/model/Allergen.kt
@@ -1,0 +1,5 @@
+package com.rafeng.bakery.improve.business.model
+
+enum class Allergen {
+    WALNUT, STRAWBERRY, MUSHROOM
+}
