+++ src/main/kotlin/com/rafeng/bakery/improve/business/model/Filling.kt
@@ -1,0 +1,8 @@
+package com.rafeng.bakery.improve.business.model
+
+data class Filling(
+    val name: String,
+    val description: String,
+    val type: FillingType,
+    val taste: Taste
+)
