+++ src/main/kotlin/com/rafeng/bakery/improve/business/model/FillingType.kt
@@ -1,0 +1,5 @@
+package com.rafeng.bakery.improve.business.model
+
+enum class FillingType {
+    JAM, LIQUID, SOLID
+}
