+++ src/main/kotlin/com/rafeng/bakery/improve/business/model/ItemFilling.kt
@@ -1,0 +1,6 @@
+package com.rafeng.bakery.improve.business.model
+
+data class ItemFilling(
+    val filling: Filling,
+    val weight: Double
+)
