+++ src/main/kotlin/com/rafeng/bakery/improve/business/model/ItemSize.kt
@@ -1,0 +1,5 @@
+package com.rafeng.bakery.improve.business.model
+
+enum class ItemSize {
+    SMALL, MEDIUM, LARGE
+}
