+++ src/main/kotlin/com/rafeng/bakery/improve/business/model/ItemSmell.kt
@@ -1,0 +1,5 @@
+package com.rafeng.bakery.improve.business.model
+
+enum class ItemSmell {
+    WEAK, MILD, STRONG,
+}
