+++ src/main/kotlin/com/rafeng/bakery/improve/business/model/ItemTopping.kt
@@ -1,0 +1,6 @@
+package com.rafeng.bakery.improve.business.model
+
+data class ItemTopping(
+    val topping: Topping,
+    val weight: Double
+)
