+++ src/main/kotlin/com/rafeng/bakery/improve/business/model/Order.kt
@@ -1,6 +1,6 @@
 package com.rafeng.bakery.improve.business.model
 
 data class Order(
-    val sellItems: List<SellItem>,
-    val total: Double
+    val sellItems: MutableList<SellItem>,
+    var total: Double
 )
