+++ src/main/kotlin/com/rafeng/bakery/improve/business/model/Recipe.kt
@@ -1,0 +1,9 @@
+package com.rafeng.bakery.improve.business.model
+
+
+data class Recipe(
+    val name: String,
+    val description: String,
+    val expirationPeriod: Int,
+    val cookingTime: Double
+)
