+++ src/main/kotlin/com/rafeng/bakery/improve/business/model/Taste.kt
@@ -1,0 +1,5 @@
+package com.rafeng.bakery.improve.business.model
+
+enum class Taste {
+    SWEET, SPICY, SALTY, SOUR
+}
