+++ src/main/kotlin/com/rafeng/bakery/improve/business/model/Topping.kt
@@ -1,0 +1,8 @@
+package com.rafeng.bakery.improve.business.model
+
+data class Topping(
+    val name: String,
+    val description: String,
+    val type: ToppingType,
+    val taste: Taste
+)
