+++ src/main/kotlin/com/rafeng/bakery/improve/business/model/ToppingType.kt
@@ -1,0 +1,5 @@
+package com.rafeng.bakery.improve.business.model
+
+enum class ToppingType {
+    CREAM, SHAVING, LIQUID
+}
