+++ src/main/kotlin/com/rafengimprove/bakerybusiness/model/Item.kt
@@ -1,0 +1,14 @@
+package com.rafengimprove.bakerybusiness.model
+
+//open class Item(
+//    val name: String,
+//    val weight: Double,
+//    val size: ItemSize,
+//    val smell: ItemSmell,
+//    val taste: Taste,
+//    val filling: List<ItemFilling>,
+//    val topping: List<ItemTopping>,
+//    val calories: Int,
+//    val allergens: List<Allergen>,
+//    val isVegan: Boolean
+//)
